# docs/assets/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LimitesIn": {
        "properties": {
          "acao_ao_estourar": {
            "pattern": "^(degradar|alertar|bloquear)$",
            "title": "Acao Ao Estourar",
            "type": "string"
          },
          "alerta_pct": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Alerta Pct",
            "type": "integer"
          },
          "cotacao_brl": {
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Cotacao Brl",
            "type": "number"
          },
          "orcamento_mensal_usd": {
            "minimum": 0.0,
            "title": "Orcamento Mensal Usd",
            "type": "number"
          },
          "teto_tokens_dia": {
            "minimum": 0.0,
            "title": "Teto Tokens Dia",
            "type": "integer"
          }
        },
        "required": [
          "orcamento_mensal_usd",
          "teto_tokens_dia",
          "alerta_pct",
          "acao_ao_estourar",
          "cotacao_brl"
        ],
        "title": "LimitesIn",
        "type": "object"
      },
...
      "PrecoIn": {
        "properties": {
          "cache_escrita": {
            "default": 0,
            "minimum": 0.0,
            "title": "Cache Escrita",
            "type": "number"
          },
          "cache_leitura": {
            "default": 0,
            "minimum": 0.0,
            "title": "Cache Leitura",
            "type": "number"
          },
          "entrada": {
            "minimum": 0.0,
            "title": "Entrada",
            "type": "number"
          },
          "saida": {
            "minimum": 0.0,
            "title": "Saida",
            "type": "number"
          }
        },
        "required": [
          "entrada",
          "saida"
        ],
        "title": "PrecoIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "api__routers__corretores__CorretorIn": {
        "properties": {
          "ativo": {
            "default": true,
            "title": "Ativo",
            "type": "boolean"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "foto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foto"
          },
          "nome": {
            "minLength": 2,
            "title": "Nome",
            "type": "string"
          },
          "regioes": {
            "items": {
              "type": "string"
            },
            "title": "Regioes",
            "type": "array"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          }
        },
        "required": [
          "nome"
        ],
        "title": "CorretorIn",
        "type": "object"
      },
      "api__routers__leads__CorretorIn": {
        "properties": {
          "corretor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corretor Id"
          }
        },
        "title": "CorretorIn",
        "type": "object"
      }
    },
    "securitySchemes": {
      "Token do painel": {
        "description": "Perfil local: o valor de `SDR_PAINEL_TOKEN` (em desenvolvimento, `dev-token`). Perfil AWS: o JWT emitido pelo Cognito.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Projeto Mora — Vértice Imóveis (POC acadêmica)"
    },
    "description": "\nAPI do **Mora**, o agente SDR imobiliário da Vértice Imóveis.\n\nA etiqueta de cada rota diz quem a usa e o que ela exige:\n\n| Etiqueta | Quem usa | Autenticação |\n|---|---|---|\n| `público` | site vitrine e chat | nenhuma |\n| `corretor` | painel do corretor | `Authorization: Bearer …` |\n| `operacao` | rotinas de atendimento no painel | `Authorization: Bearer …` |\n| `admin` | configuração, governança de IA e auditoria | `Authorization: Bearer …` |\n\n**Como autenticar no perfil local:** clique em **Authorize** e informe o valor de\n`SDR_PAINEL_TOKEN` (no ambiente de desenvolvimento, `dev-token`). No perfil AWS o token é o\nJWT do Cognito, validado pelo authorizer do API Gateway antes de chegar aqui.\n\nToda alteração feita por estas rotas é registrada em auditoria (ver `GET /auditoria`).\n",
    "license": {
      "name": "Uso acadêmico — licença a confirmar"
    },
    "title": "Mora — SDR Imobiliário API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/corretores": {
      "get": {
        "operationId": "listar_corretores_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Listar",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "criar_corretores_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api__routers__corretores__CorretorIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Criar",
        "tags": [
          "admin"
        ]
      }
    },
    "/corretores/{corretor_id}": {
      "delete": {
        "description": "Desliga o corretor: `ativo = false` e a carteira vai para quem você indicar.\n\nNão apaga o cadastro. Antes, o DELETE removia a linha e os leads ficavam apontando para um id\ninexistente — em handoff, atribuídos a ninguém e fora da lista de todo mundo. Desativar mantém\no histórico (quem atendeu quem) e tira o corretor do roteamento de novos leads.",
        "operationId": "desativar_corretores__corretor_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "corretor_id",
            "required": true,
            "schema": {
              "title": "Corretor Id",
              "type": "string"
            }
          },
          {
            "description": "Para onde vai a carteira: id de outro corretor, `equipe` (fila sem dono) ou `auto` (distribui pelo roteamento por região e carga). Obrigatório quando a carteira não está vazia.",
            "in": "query",
            "name": "destino",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Para onde vai a carteira: id de outro corretor, `equipe` (fila sem dono) ou `auto` (distribui pelo roteamento por região e carga). Obrigatório quando a carteira não está vazia.",
              "title": "Destino"
            }
          },
          {
            "description": "Apaga o cadastro em vez de desativar. Só permitido com a carteira vazia — existe para o cadastro criado por engano, não para desligar quem já atendeu.",
            "in": "query",
            "name": "remover_cadastro",
            "required": false,
            "schema": {
              "default": false,
              "description": "Apaga o cadastro em vez de desativar. Só permitido com a carteira vazia — existe para o cadastro criado por engano, não para desligar quem já atendeu.",
              "title": "Remover Cadastro",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Desativar",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "atualizar_corretores__corretor_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "corretor_id",
            "required": true,
            "schema": {
              "title": "Corretor Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api__routers__corretores__CorretorIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Atualizar",
        "tags": [
          "admin"
        ]
      }
    },
    "/corretores/{corretor_id}/carteira": {
      "get": {
        "description": "Quantos leads abertos e visitas futuras estão no nome deste corretor.\n\nA tela chama isto antes de desligar alguém, para perguntar para onde vai a carteira.",
        "operationId": "carteira_corretores__corretor_id__carteira_get",
        "parameters": [
          {
            "in": "path",
            "name": "corretor_id",
            "required": true,
            "schema": {
              "title": "Corretor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Carteira",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/governanca/limites": {
      "get": {
        "operationId": "limites_governanca_limites_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Limites",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "salvar_limites_governanca_limites_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LimitesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Salvar Limites",
        "tags": [
          "admin"
        ]
      }
    },
    "/governanca/precos": {
      "get": {
        "description": "Preços por 1M de tokens: (entrada, saída, escrita de cache, leitura de cache) em USD.",
        "operationId": "precos_governanca_precos_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Precos",
        "tags": [
          "admin"
        ]
      }
    },
    "/governanca/precos/{modelo}": {
      "delete": {
        "operationId": "restaurar_preco_governanca_precos__modelo__delete",
        "parameters": [
          {
            "in": "path",
            "name": "modelo",
            "required": true,
            "schema": {
              "title": "Modelo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Restaurar Preco",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "salvar_preco_governanca_precos__modelo__put",
        "parameters": [
          {
            "in": "path",
            "name": "modelo",
            "required": true,
            "schema": {
              "title": "Modelo",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrecoIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Salvar Preco",
        "tags": [
          "admin"
        ]
      }
    },
    "/governanca/uso": {
      "get": {
        "description": "Consumo do período com variação, série diária, quebra por modelo/nó/papel e chamadas recentes.",
        "operationId": "uso_governanca_uso_get",
        "parameters": [
          {
            "in": "query",
            "name": "dias",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 180,
              "minimum": 1,
              "title": "Dias",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Uso",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/leads/{lead_id}/corretor": {
      "put": {
        "description": "Atribui (ou troca/remove) o corretor responsável pelo lead; visitas futuras acompanham.",
        "operationId": "atribuir_leads__lead_id__corretor_put",
        "parameters": [
          {
            "in": "path",
            "name": "lead_id",
            "required": true,
            "schema": {
              "title": "Lead Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api__routers__leads__CorretorIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Atribuir",
        "tags": [
          "corretor"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Perfil local (docker compose)",
      "url": "http://localhost:8000"
    },
    {
      "description": "Perfil AWS — a confirmar no deploy",
      "url": "https://api.exemplo.com.br"
    }
  ],
...
}